breytti file structure, lagaði smá css og breytti í typescript
