+++ home/index.html
@@ -4,8 +4,8 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Costco</title>
-    <link href="./style.css" type="text/css" rel="stylesheet" />
-    <link />
+    <link href="../style.css" type="text/css" rel="stylesheet" />
+    <link rel="stylesheet" href="./homeMainContent.css">
   </head>
   <body>
     <!-- Header Section -->
@@ -13,18 +13,18 @@
       <div class="nav-bar flex-x">
         <div>
           <a href="/" class="logo"
-            ><img src="./images/Costco-Logo.wine.svg"
+            ><img src="../images/Costco-Logo.wine.svg"
           /></a>
         </div>
         <!-- Navigation Icons (search, user profile, wishlist, cart, hamburger menu) -->
         <div class="nav-bar-icons">
-          <a href="/"> <img src="./images/magnifying-glass-solid.svg" /></a>
-          <a href="/"> <img src="./images/user-regular.svg" /></a>
-          <a href="/"> <img src="./images/heart-regular.svg" /></a>
-          <a href="/"> <img src="./images/cart-shopping-solid.svg" /></a>
+          <a href="/"> <img src="../images/magnifying-glass-solid.svg" /></a>
+          <a href="/"> <img src="../images/user-regular.svg" /></a>
+          <a href="/"> <img src="../images/heart-regular.svg" /></a>
+          <a href="/"> <img src="../images/cart-shopping-solid.svg" /></a>
           <!-- Hamburger Menu for mobile view -->
           <img
-            src="images/hamburger.svg"
+            src="../images/hamburger.svg"
             alt="hamburger icon"
             class="hamburger"
           />
@@ -44,7 +44,7 @@
               <!-- Subcategories -->
               <li class="matur"><a href="#">Matur</a></li>
               <ul class="mobile-links flokkar">
-                <li><a href="/bananar.html">Ávextir og grænmeti</a></li>
+                <li><a href="../vorusida/vorusida.html">Ávextir og grænmeti</a></li>
                 <li><a href="">Fiskur</a></li>
                 <li><a href="">Kjöt</a></li>
                 <li><a href="">Mjólkurvörur og egg</a></li>
@@ -76,8 +76,8 @@
 
     <!-- Opening Hours Section -->
     <div class="opening-hours">
-      <a href="opnun.html"
-        ><img src="images/clock-regular.svg" /><span class="blacklink"
+      <a href="../opnun/opnun.html"
+        ><img src="../images/clock-regular.svg" /><span class="blacklink"
           >Opnunartími</span
         ></a
       >
@@ -87,8 +87,8 @@
     <div class="app-downloads">
       <p>Ná í Costco appið</p>
       <div class="app-icons">
-        <img src="./images/apple-brands-solid.svg" alt="Apple icon" />
-        <img src="./images/google-play-brands-solid.svg" alt="play icon" />
+        <img src="../images/apple-brands-solid.svg" alt="Apple icon" />
+        <img src="../images/google-play-brands-solid.svg" alt="play icon" />
       </div>
     </div>
 
@@ -111,25 +111,25 @@
         </div>
 
         <div class="card-container">
-          <div class="card card-1">
+          <div class="card categories-card">
             <img
-              src="./images/article_291139_the-top-10-healthiest-foods-for-kids_-02-4b745e57928c4786a61b47d8ba920058.jpg"
+              src="../images/article_291139_the-top-10-healthiest-foods-for-kids_-02-4b745e57928c4786a61b47d8ba920058.jpg"
               alt="Matur"
             />
             <div class="card-text">Matur</div>
             <div class="card-icon">
-              <img src="/images/Vector.svg" alt="vector" />
+              <img src="../images/Vector.svg" alt="vector" />
             </div>
           </div>
 
-          <div class="card card-2">
+          <div class="card categories-card">
             <img
-              src="./images/Screenshot 2024-10-08 at 12.36.20.png"
+              src="../images/Screenshot 2024-10-08 at 12.36.20.png"
               alt="Fatnaður"
             />
             <div class="card-text">Fatnaður</div>
             <div class="card-icon">
-              <img src="/images/Vector1.svg" alt="vector1" />
+              <img src="../images/Vector1.svg" alt="vector1" />
             </div>
           </div>
         </div>
@@ -144,7 +144,7 @@
           </div>
           <div class="card-container">
             <div class="card-5">
-              <img src="./images/costco-mobile-app.webp" alt="mobile-app" />
+              <img src="../images/costco-mobile-app.webp" alt="mobile-app" />
             </div>
           </div>
           <!-- Information about downloading the app -->
@@ -171,19 +171,19 @@
         </div>
 
         <div class="card-container">
-          <div class="card card-1">
-            <img src="./images/sjon.png" alt="Vín" />
+          <div class="card categories-card">
+            <img src="../images/sjon.png" alt="Vín" />
             <div class="card-text">Sjón</div>
             <div class="card-icon">
-              <img src="/images/Vector2.svg" alt="vector1" />
+              <img src="../images/Vector2.svg" alt="vector1" />
             </div>
           </div>
 
-          <div class="card card-2">
-            <img src="./images/vinbud.png" alt="Vín" />
+          <div class="card categories-card">
+            <img src="../images/vinbud.png" alt="Vín" />
             <div class="card-text">Vín</div>
             <div class="card-icon">
-              <img src="/images/Vector3.svg" alt="vector1" />
+              <img src="../images/Vector3.svg" alt="vector1" />
             </div>
           </div>
         </div>
@@ -198,11 +198,11 @@
 
         <div class="card-container">
           <div class="card-col1 card-3">
-            <img src="./images/netklubbur.png" alt="netklúbbur" />
+            <img src="../images/netklubbur.png" alt="netklúbbur" />
           </div>
 
           <div class="card-col card-4">
-            <img src="./images/netklubbur-simi.png" alt="sími" />
+            <img src="../images/netklubbur-simi.png" alt="sími" />
           </div>
         </div>
 
@@ -227,17 +227,17 @@
       <div class="container">
         <div class="uppl">
           <a href="#">
-            <img src="images/house-solid.svg" />
+            <img src="../images/house-solid.svg" />
             <p>Costco Wholesale Iceland, Kauptún 3, 210 Garðabær</p></a
           >
-          <a href="#"> <img src="images/at-solid.svg" />costco@costco.is</a>
-          <a href="#"> <img src="images/phone-solid.svg" />532 5555</a>
+          <a href="#"> <img src="../images/at-solid.svg" />costco@costco.is</a>
+          <a href="#"> <img src="../images/phone-solid.svg" />532 5555</a>
         </div>
 
         <div class="socials">
-          <a href="#"> <img src="images/facebook-brands-solid.svg" /></a>
-          <a href="#"> <img src="images/instagram-brands-solid.svg" /></a>
-          <a href="#"> <img src="images/youtube-brands-solid (1).svg" /></a>
+          <a href="#"> <img src="../images/facebook-brands-solid.svg" /></a>
+          <a href="#"> <img src="../images/instagram-brands-solid.svg" /></a>
+          <a href="#"> <img src="../images/youtube-brands-solid (1).svg" /></a>
         </div>
       </div>
       <p class="copyright">
@@ -247,6 +247,7 @@
     </footer>
 
     <!-- Load the script file for the menu overlay -->
-    <script src="./main.js"></script>
+    <script src="../main.js"></script>
+    <script src="../header.js"></script>
   </body>
 </html>
