+++ opnun/opnun.css
@@ -1,0 +1,34 @@
+/* ======================
+       OPNUNARTÍMI
+   ======================*/
+
+   .opnunh2 {
+    font-size: 24px;
+    font-weight: bold;
+    display: flex;
+    justify-content: center;
+    margin-bottom: 16px;
+  }
+  
+  .opnun-container {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-end;
+    margin-left: 20%;
+    margin-right: 20%;
+    font-size: 20px;
+    margin-bottom: 30px;
+  }
+  .opnun-container p {
+    margin-top: 10px;
+  }
+  .opnunartimar-container {
+    margin: 28px 0 76px 0;
+  }
+  
+  @media (max-width: 860px) {
+    .opnun-container {
+      margin-left: 15%;
+      margin-right: 15%;
+    }
+  }
