+++ opnun/opnun.html
@@ -4,23 +4,24 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Document</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="../style.css">
+    <link rel="stylesheet" href="./opnun.css">
   </head>
   <body>
     <header>
       <div class="nav-bar flex-x">
         <div>
-          <a href="/" class="logo"
-            ><img src="./images/Costco-Logo.wine.svg"
+          <a href="../home/index.html" class="logo"
+            ><img src="../images/Costco-Logo.wine.svg"
           /></a>
         </div>
         <div class="nav-bar-icons">
-          <a href="/"> <img src="./images/magnifying-glass-solid.svg" /></a>
-          <a href="/"> <img src="./images/user-regular.svg" /></a>
-          <a href="/"> <img src="./images/heart-regular.svg" /></a>
-          <a href="/"> <img src="./images/cart-shopping-solid.svg" /></a>
+          <a href="/"> <img src="../images/magnifying-glass-solid.svg" /></a>
+          <a href="/"> <img src="../images/user-regular.svg" /></a>
+          <a href="/"> <img src="../images/heart-regular.svg" /></a>
+          <a href="/"> <img src="../images/cart-shopping-solid.svg" /></a>
           <img
-            src="images/hamburger.svg"
+            src="../images/hamburger.svg"
             alt="hamburger icon"
             class="hamburger"
           />
@@ -39,7 +40,7 @@
             <div class="underlist">
               <li class="matur"><a href="#">Matur</a></li>
               <ul class="mobile-links flokkar">
-                <li><a href="/bananar.html">Ávextir og grænmeti</a></li>
+                <li><a href="../vorusida/vorusida.html">Ávextir og grænmeti</a></li>
                 <li><a href="">Fiskur</a></li>
                 <li><a href="">Kjöt</a></li>
                 <li><a href="">Mjólkurvörur og egg</a></li>
@@ -178,17 +179,17 @@
       <div class="container">
         <div class="uppl">
           <a href="#">
-            <img src="images/house-solid.svg" />
+            <img src="../images/house-solid.svg" />
             <p>Costco Wholesale Iceland, Kauptún 3, 210 Garðabær</p></a
           >
-          <a href="#"> <img src="images/at-solid.svg" />costco@costco.is</a>
-          <a href="#"> <img src="images/phone-solid.svg" />532 5555</a>
+          <a href="#"> <img src="../images/at-solid.svg" />costco@costco.is</a>
+          <a href="#"> <img src="../images/phone-solid.svg" />532 5555</a>
         </div>
 
         <div class="socials">
-          <a href="#"> <img src="images/facebook-brands-solid.svg" /></a>
-          <a href="#"> <img src="images/instagram-brands-solid.svg" /></a>
-          <a href="#"> <img src="images/youtube-brands-solid (1).svg" /></a>
+          <a href="#"> <img src="../images/facebook-brands-solid.svg" /></a>
+          <a href="#"> <img src="../images/instagram-brands-solid.svg" /></a>
+          <a href="#"> <img src="../images/youtube-brands-solid (1).svg" /></a>
         </div>
       </div>
       <p class="copyright">
@@ -197,6 +198,7 @@
       </p>
     </footer>
 
-    <script src="main.js"></script>
+    <script src="../main.js"></script>
+    <script src="../header.js"></script>
   </body>
 </html>
