+++ vorusida/vorusida.css
@@ -1,0 +1,68 @@
+/* ======================
+       Product page 
+   ======================*/
+
+/* Matur - ávextir og grænmeti */
+.food {
+  display: flex;
+  justify-content: flex-start;
+  padding-left: 28px;
+  align-items: center;
+  margin-bottom: 14px;
+  font-family: futura;
+  gap: 10px;
+}
+#matur {
+  text-decoration: underline;
+  color: black;
+  font-weight: 300;
+  font-size: 18px;
+}
+#matur:hover{
+  cursor: pointer;
+}
+#matur:active{
+  color: #aaa;
+}
+.food p {
+  font-size: 14px;
+  font-weight: 500;
+}
+.rule {
+  width: 15px;
+  height: 1px;
+  background-color: black;
+  margin-right: 10px;
+}
+.rule::after {
+  content: "";
+  height: 5px;
+  width: 5px;
+  border-radius: 10px;
+  background-color: black;
+  color: black;
+  float: right;
+  margin-top: -2px;
+}
+
+/* buttons í vörum! */
+
+.btn-container {
+  display: flex;
+  justify-content: flex-start;
+  padding-left: 20px;
+  gap: 10px;
+}
+button {
+  background-color: #0073b9;
+  padding: 5px 20px;
+  color: white;
+  border-radius: 20px;
+  font-size: 14px;
+  cursor: pointer;
+  font-family: futura;
+  border: 2px solid #0073b9;
+}
+.active {
+  display: block;
+}
